bug fixes and added damage and projectiles
i added both player and enemy taking damage. the player is able to shoot projectiles and the enemy hits the player if the player is in melee range of enemy. enemy has 3 health and player has 5 (adjustable). attack damage of both is also adjustable. can see hit detection and player/enemy health in the debugging console. fixed a few collision bugs as well.