Updated coins
Made it so that there is a total accumulation of coins. Display is shown at all times with the total, with a display below that will only appear if coins are being collected. there is the variable "cumulative_coin_total" that can be used for later vendor, buying and selling uses, holding the total amount of coins the player has.