Created Center Objective
made a temp center objective so that the enemy will choose between it and the player, depending on what is closer. if the enemy reaches the center objective it will stick to it, even if the player is "closer" making the center objective a priority over the player. have not yet implemented a health system for the objective but got the main functions for it down